html/semantics/embedded-content-0/media-elements/currentTime.html

<!doctype html>
<title>currentTime is intially zero</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
test(function() {
  var v = document.createElement('video');
  assert_equals(v.currentTime, 0);
});
</script>
